Refactor to follow modular monolith 2
