+++ src/main/java/enterprise/hoversprite/modules/order/IOrderService.java
@@ -1,0 +1,4 @@
+package enterprise.hoversprite.modules.order;
+
+public interface IOrderService {
+}
