+++ src/main/java/enterprise/hoversprite/modules/order/OrderController.java
@@ -1,6 +1,5 @@
-package enterprise.hoversprite.modules.order.controller;
+package enterprise.hoversprite.modules.order;
 
-import enterprise.hoversprite.modules.order.repository.OrderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
