+++ src/main/java/enterprise/hoversprite/modules/order/OrderRepository.java
@@ -1,0 +1,8 @@
+package enterprise.hoversprite.modules.order;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+interface OrderRepository extends JpaRepository<Order, Long> {
+}
