+++ src/main/java/enterprise/hoversprite/modules/order/OrderService.java
@@ -1,0 +1,11 @@
+package enterprise.hoversprite.modules.order;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+class OrderService implements IOrderService {
+
+    @Autowired
+    private OrderRepository orderRepository;
+}
