--- src/main/java/enterprise/hoversprite/modules/order/repository/OrderRepository.java
@@ -1,9 +1,0 @@
-package enterprise.hoversprite.modules.order.repository;
-
-import enterprise.hoversprite.modules.order.model.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface OrderRepository extends JpaRepository<Order, Long> {
-}
